requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
google-generativeai
numpy
python-dotenv
tqdm  # Добавил для красивого индикатора прогресса в indexer.py